+++ src/main/kotlin/org/pcsoft/intellij/plugin/inno_setup/language/parser/preprocessor/expression/IsPreprocessorExprToken.kt
@@ -21,6 +21,8 @@ enum class IsPreprocessorExprTokenType {
     OPERATOR,
     LPAREN,
     RPAREN,
+    LBRACKET,
+    RBRACKET,
     COMMA,
     QUESTION,
     COLON,